refactor: centralize ecosystem dependency config
