--- services/py/embedding_service/ecosystem.dependencies.js
@@ -1,3 +1,0 @@
-export default {
-  apps: [],
-};
