+++ services/py/stt/ecosystem.config.js
@@ -1,30 +1,28 @@
-import path from "path";
-import { fileURLToPath } from "url";
-import { defineApp } from "../../../dev/pm2Helpers.js";
-import deps from "./ecosystem.dependencies.js";
+import path from 'path';
+import { fileURLToPath } from 'url';
+import { defineApp } from '../../../dev/pm2Helpers.js';
+import deps from '../../shared/ecosystem.dependencies.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
-const root = path.join(__dirname, "../../..");
+const root = path.join(__dirname, '../../..');
 
 if (!process.env.PROMETHEAN_ROOT_ECOSYSTEM) {
-  defineApp.PYTHONPATH = root;
-  defineApp.HEARTBEAT_PORT = 5005;
+    defineApp.PYTHONPATH = root;
+    defineApp.HEARTBEAT_PORT = 5005;
 }
 
 const apps = [
-  defineApp("stt", "uv", ["run", "python", "-m", "service"], {
-    cwd: __dirname,
-    watch: [__dirname],
-    env: {
-      FLASK_APP: "app.py",
-      FLASK_ENV: "production",
-    },
-  }),
+    defineApp('stt', 'uv', ['run', 'python', '-m', 'service'], {
+        cwd: __dirname,
+        watch: [__dirname],
+        env: {
+            FLASK_APP: 'app.py',
+            FLASK_ENV: 'production',
+        },
+    }),
 ];
 
-const allApps = !process.env.PROMETHEAN_ROOT_ECOSYSTEM
-  ? [...apps, ...(deps?.apps || [])]
-  : apps;
+const allApps = !process.env.PROMETHEAN_ROOT_ECOSYSTEM ? [...apps, ...(deps?.apps || [])] : apps;
 
 export default { apps: allApps };
