+++ services/shared/ecosystem.dependencies.js
@@ -1,3 +1,3 @@
 export default {
-  apps: [],
+    apps: [],
 };
